# openapi.yaml
openapi: 3.1.0
info:
  title: Issuer Service API
  description: Generic swiyu Issuer Service service
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /.well-known/openid-credential-issuer/{tenantId}:
    get:
      tags:
      - Well-known endpoints API
      summary: Information about credentials which can be issued.
      operationId: getIssuerMetadataByTenantId
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Credential Issuer Metadata
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /.well-known/openid-credential-issuer/oid4vci/{tenantId}:
    get:
      tags:
      - Well-known endpoints API
      summary: Information about credentials which can be issued.
      operationId: getIssuerMetadataByTenantId_1
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Credential Issuer Metadata
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /{tenantId}/.well-known/openid-credential-issuer:
    get:
      tags:
      - Well-known endpoints API
      summary: Information about credentials which can be issued.
      operationId: getIssuerMetadataByTenantId_2
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Credential Issuer Metadata
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /oid4vci/{tenantId}/.well-known/openid-credential-issuer:
    get:
      tags:
      - Well-known endpoints API
      summary: Information about credentials which can be issued.
      operationId: getIssuerMetadataByTenantId_3
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Credential Issuer Metadata
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /.well-known/openid-credential-issuer/oid4vci:
    get:
      tags:
      - Well-known endpoints API
      summary: Information about credentials which can be issued.
      operationId: getIssuerMetadata
      parameters:
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Credential Issuer Metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuerMetadata"
            application/jwt:
              schema:
                type: string
                description: Signed issuer metadata as JWT
        "406":
          description: "None of the requested Accept media types are supported, or\
            \ signed metadata was requested but is not configured for this issuer"
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /oid4vci/.well-known/openid-credential-issuer:
    get:
      tags:
      - Well-known endpoints API
      summary: Information about credentials which can be issued.
      operationId: getIssuerMetadata_1
      parameters:
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Credential Issuer Metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuerMetadata"
            application/jwt:
              schema:
                type: string
                description: Signed issuer metadata as JWT
        "406":
          description: "None of the requested Accept media types are supported, or\
            \ signed metadata was requested but is not configured for this issuer"
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /.well-known/openid-credential-issuer:
    get:
      tags:
      - Well-known endpoints API
      summary: Information about credentials which can be issued.
      operationId: getIssuerMetadata_2
      parameters:
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Credential Issuer Metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuerMetadata"
            application/jwt:
              schema:
                type: string
                description: Signed issuer metadata as JWT
        "406":
          description: "None of the requested Accept media types are supported, or\
            \ signed metadata was requested but is not configured for this issuer"
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /oid4vci/{tenantId}/.well-known/oauth-authorization-server:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve tenant-specific OAuth 2.0 Authorization Server Metadata
      description: "Returns the Authorization Server configuration metadata for the\
        \ given tenant in accordance with RFC 8414. Depending on the 'Accept' header,\
        \ the response is provided either as an unsigned JSON document or as a signed\
        \ JWT. The metadata includes issuer information, endpoint URLs (e.g., token\
        \ endpoint), supported grant types and extensions required for OpenID for\
        \ Verifiable Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadataByTenantId
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /{tenantId}/.well-known/oauth-authorization-server/oid4vci:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve tenant-specific OAuth 2.0 Authorization Server Metadata
      description: "Returns the Authorization Server configuration metadata for the\
        \ given tenant in accordance with RFC 8414. Depending on the 'Accept' header,\
        \ the response is provided either as an unsigned JSON document or as a signed\
        \ JWT. The metadata includes issuer information, endpoint URLs (e.g., token\
        \ endpoint), supported grant types and extensions required for OpenID for\
        \ Verifiable Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadataByTenantId_1
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /oid4vci/{tenantId}/.well-known/openid-configuration:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve tenant-specific OAuth 2.0 Authorization Server Metadata
      description: "Returns the Authorization Server configuration metadata for the\
        \ given tenant in accordance with RFC 8414. Depending on the 'Accept' header,\
        \ the response is provided either as an unsigned JSON document or as a signed\
        \ JWT. The metadata includes issuer information, endpoint URLs (e.g., token\
        \ endpoint), supported grant types and extensions required for OpenID for\
        \ Verifiable Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadataByTenantId_2
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /{tenantId}/.well-known/oauth-authorization-server:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve tenant-specific OAuth 2.0 Authorization Server Metadata
      description: "Returns the Authorization Server configuration metadata for the\
        \ given tenant in accordance with RFC 8414. Depending on the 'Accept' header,\
        \ the response is provided either as an unsigned JSON document or as a signed\
        \ JWT. The metadata includes issuer information, endpoint URLs (e.g., token\
        \ endpoint), supported grant types and extensions required for OpenID for\
        \ Verifiable Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadataByTenantId_3
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /.well-known/oauth-authorization-server/{tenantId}:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve tenant-specific OAuth 2.0 Authorization Server Metadata
      description: "Returns the Authorization Server configuration metadata for the\
        \ given tenant in accordance with RFC 8414. Depending on the 'Accept' header,\
        \ the response is provided either as an unsigned JSON document or as a signed\
        \ JWT. The metadata includes issuer information, endpoint URLs (e.g., token\
        \ endpoint), supported grant types and extensions required for OpenID for\
        \ Verifiable Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadataByTenantId_4
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /{tenantId}/.well-known/openid-configuration:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve tenant-specific OAuth 2.0 Authorization Server Metadata
      description: "Returns the Authorization Server configuration metadata for the\
        \ given tenant in accordance with RFC 8414. Depending on the 'Accept' header,\
        \ the response is provided either as an unsigned JSON document or as a signed\
        \ JWT. The metadata includes issuer information, endpoint URLs (e.g., token\
        \ endpoint), supported grant types and extensions required for OpenID for\
        \ Verifiable Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadataByTenantId_5
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /.well-known/oauth-authorization-server/{tenantId}/oid4vci:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve tenant-specific OAuth 2.0 Authorization Server Metadata
      description: "Returns the Authorization Server configuration metadata for the\
        \ given tenant in accordance with RFC 8414. Depending on the 'Accept' header,\
        \ the response is provided either as an unsigned JSON document or as a signed\
        \ JWT. The metadata includes issuer information, endpoint URLs (e.g., token\
        \ endpoint), supported grant types and extensions required for OpenID for\
        \ Verifiable Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadataByTenantId_6
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: No credential offer exists for the given tenant id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /.well-known/openid-configuration:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve OAuth 2.0 Authorization Server Metadata
      description: "Returns the configuration metadata of the Authorization Server\
        \ in accordance with RFC 8414. This includes URLs to endpoints (e.g., token\
        \ endpoint), supported grant types, as well as extensions for OpenID for Verifiable\
        \ Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadata
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthAuthorizationServerMetadata"
        "406":
          description: None of the requested Accept media types are supported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthAuthorizationServerMetadata"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /.well-known/oauth-authorization-server/oid4vci:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve OAuth 2.0 Authorization Server Metadata
      description: "Returns the configuration metadata of the Authorization Server\
        \ in accordance with RFC 8414. This includes URLs to endpoints (e.g., token\
        \ endpoint), supported grant types, as well as extensions for OpenID for Verifiable\
        \ Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadata_1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthAuthorizationServerMetadata"
        "406":
          description: None of the requested Accept media types are supported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthAuthorizationServerMetadata"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /.well-known/oauth-authorization-server:
    get:
      tags:
      - Well-known endpoints API
      summary: Retrieve OAuth 2.0 Authorization Server Metadata
      description: "Returns the configuration metadata of the Authorization Server\
        \ in accordance with RFC 8414. This includes URLs to endpoints (e.g., token\
        \ endpoint), supported grant types, as well as extensions for OpenID for Verifiable\
        \ Credential Issuance (OID4VCI) and DPoP."
      operationId: getAuthorizationServerMetadata_2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthAuthorizationServerMetadata"
        "406":
          description: None of the requested Accept media types are supported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthAuthorizationServerMetadata"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    ApiError:
      type: object
      description: Error response object
      properties:
        error:
          type: string
        error_description:
          type: string
        detail:
          type: string
        trace_id:
          type: string
      required:
      - error
...
    OAuthAuthorizationServerMetadata:
      type: object
      properties:
        issuer:
          type: string
          description: The Issuer Identifier
        token_endpoint:
          type: string
          description: URL of the OAuth 2.0 Token Endpoint
        dpop_signing_alg_values_supported:
          type: array
          items:
            type: string
        profile_version:
          type: string
        pre-authorized_grant_anonymous_access_supported:
          type: boolean
      required:
      - issuer
      - token_endpoint
...
  securitySchemes:
    bearer-jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT